Rename PageReader to DocumentReaderFacade and cut into it two parts: BooksPageReader and BookPropertiesPageReader
